fix: address code-review findings on RocksDbReadOps split
Two real issues surfaced by review, both fixed:

- SecondaryDB implementing RocksDbReadOps (which had CF-scoped get/
  keyMayExist/newIterator/getProperty) newly exposed 9 CF methods it
  has no legitimate way to support: RocksDB.openSecondary() has no
  multi-CF overload and SecondaryDB has no createColumnFamily, so any
  ColumnFamilyHandle passed in is guaranteed to come from a different
  rocksdb_t* -- undefined behavior at the C++ level, up to a JVM crash.
  Split the CF-scoped overloads into a new RocksDbCfReadOps extends
  RocksDbReadOps. Only ReadWriteDB/TtlDB/BlobDB (via RocksDbWriteOps)
  and ReadOnlyDB implement it -- both genuinely support multi-CF open.
  SecondaryDB now implements plain RocksDbReadOps and has zero CF
  methods again, matching its actual capability.
- defaultReadOpts()/defaultWriteOpts() were abstract, forcing every one
  of the 5 implementors to keep its own field + constructor assignment
  + accessor override for what is, in every current case, the exact
  same RocksDB.DEFAULT_READ_OPTIONS/DEFAULT_WRITE_OPTIONS singleton --
  the copy-paste this PR set out to remove, one layer up. Made both
  default methods returning the shared singleton directly; the field/
  override stays overridable if a future implementor ever needs its
  own.

Two other findings (TtlDB.ingestExternalFile silently corrupting reads
due to the TTL timestamp suffix; createColumnFamily not getting TTL
semantics since rocksdb_create_column_family_with_ttl isn't mapped) are
pre-existing in the original TtlDB.java, predate this refactor, and the
ingest caveat is already documented and tested around in TtlDBTest --
left alone as out of scope.

./mvnw test and ./mvnw javadoc:javadoc -pl core both pass clean.
