+++ CLAUDE.md
@@ -183,7 +183,7 @@ instance method listed below lives on the DB type (`ReadWriteDB`, `TtlDB`, …),
 | KeyMayExist             | `ReadWriteDB.keyMayExist` — byte[], ByteBuffer, MemorySegment, ReadOptions overload                                           |
 | DB Properties           | `Property.java` (enum of well-known names); `ReadWriteDB.getProperty`, `ReadWriteDB.getLongProperty`, same on `TransactionDB`     |
 | Statistics              | `HistogramType.java`, `TickerType.java`, `StatsLevel.java`, `StatisticsHistogramData.java`                                |
-| Shared utilities        | `RocksDB.java` statics (`errHolder`, `checkError`, `toNative`), `NativeObject.java`, `NativeLibrary.java`, `MemorySize.java`, `SequenceNumber.java`, `RocksDBException.java`; `RocksDbReadOps.java`/`RocksDbWriteOps.java` — default-method interfaces sharing the direct put/get/merge/delete/iterate/flush/compaction/CF-management surface across `ReadWriteDB`, `TtlDB`, `BlobDB`, `ReadOnlyDB`, `SecondaryDB` (not `TransactionDB`/`OptimisticTransactionDB`, whose direct ops bind their own `MethodHandle`s per the no-shared-call-site rule) |
+| Shared utilities        | `RocksDB.java` statics (`errHolder`, `checkError`, `toNative`), `NativeObject.java`, `NativeLibrary.java`, `MemorySize.java`, `SequenceNumber.java`, `RocksDBException.java`; `RocksDbReadOps.java`/`RocksDbCfReadOps.java`/`RocksDbWriteOps.java` — default-method interfaces sharing the direct put/get/merge/delete/iterate/flush/compaction/CF-management surface. `RocksDbReadOps` (no CF methods) is implemented by `SecondaryDB`, which has no multi-CF open path; `RocksDbCfReadOps` (adds CF-scoped reads) by `ReadOnlyDB`; `RocksDbWriteOps` (extends `RocksDbCfReadOps`) by `ReadWriteDB`/`TtlDB`/`BlobDB`. Not implemented by `TransactionDB`/`OptimisticTransactionDB`, whose direct ops bind their own `MethodHandle`s per the no-shared-call-site rule |
 | Compaction control      | `CompactOptions.java`, `WaitForCompactOptions.java`; `ReadWriteDB.compactRange`, `suggestCompactRange`, `disableFileDeletions`, `enableFileDeletions`, `ReadWriteDB.waitForCompact` |
 | DeleteRange             | `ReadWriteDB.deleteRange` (all three tiers), `WriteBatch.deleteRange`                                                     |
 | Merge                   | `merge()` on `ReadWriteDB`, `TtlDB`, `BlobDB`, `OptimisticTransactionDB`, `TransactionDB`, `Transaction`, `WriteBatch` (all tiers + CF variants); `MergeOperator.java` (`uint64Add()` built-in, `custom(String, FullMergeFn)` callback-based), wired via `Options.setMergeOperator` |
