+++ core/src/main/java/io/github/dfa1/rocksdbffm/ReadWriteDB.java
@@ -14,28 +14,13 @@ import java.lang.foreign.MemorySegment;
 /// ```
 public final class ReadWriteDB extends NativeObject implements RocksDbWriteOps {
 
-	private final WriteOptions writeOpts;
-	private final ReadOptions readOpts;
-
 	ReadWriteDB(MemorySegment ptr) {
 		super(ptr);
-		this.writeOpts = RocksDB.DEFAULT_WRITE_OPTIONS;
-		this.readOpts = RocksDB.DEFAULT_READ_OPTIONS;
 	}
 
 	@Override
 	public MemorySegment dbPtr() {
 		return ptr();
-	}
-
-	@Override
-	public WriteOptions defaultWriteOpts() {
-		return writeOpts;
-	}
-
-	@Override
-	public ReadOptions defaultReadOpts() {
-		return readOpts;
 	}
 
 	// -----------------------------------------------------------------------
